reserva

+++ Models/Suite.cs
@@ -4,7 +4,7 @@ namespace DesafioProjetoHospedagem.Models
     {
         public Suite() { }
 
-        public Suite(string tipoSuite, int capacidade, decimal valorDiaria)
+        public Suite(string tipoSuite, int capacidade, double valorDiaria)
         {
             TipoSuite = tipoSuite;
             Capacidade = capacidade;
@@ -13,6 +13,6 @@ namespace DesafioProjetoHospedagem.Models
 
         public string TipoSuite { get; set; }
         public int Capacidade { get; set; }
-        public decimal ValorDiaria { get; set; }
+        public double ValorDiaria { get; set; }
     }
 }
